2021 March update

+++ app/services/celo/blocks_summary.rb
@@ -1,5 +1,5 @@
 class Celo::BlocksSummary < Common::Resource
-  field :block_time_avg, type: :integer, default: 0
+  field :block_time_avg, type: :float, default: 0
   field :time_bucket, type: :timestamp
 
   def total
